Replace some UShort to Short in order to fix tests

+++ src/commonMain/kotlin/lt/libredrop/peerdiscovery/PeerDiscovery.kt
@@ -9,7 +9,7 @@ import lt.libredrop.peerdiscovery.network.TransportProtocol
 import lt.libredrop.peerdiscovery.network.UUID
 import lt.libredrop.peerdiscovery.network.randomUUID
 
-class PeerDiscovery(private val networkDriver: NetworkDriver, private val port: UShort = 5530u) {
+class PeerDiscovery(private val networkDriver: NetworkDriver, private val port: Short = 5530) {
     suspend fun start(
         serviceName: String,
         uuid: UUID = randomUUID(),
@@ -27,6 +27,6 @@ class PeerDiscovery(private val networkDriver: NetworkDriver, private val port: 
 
         networkDriver.broadcast(thisPeer.createBinaryMessage(), port)
 
-        return networkDriver.listenForPeers(port)
+        return networkDriver.listenForPeers(port.toUShort())
     }
 }
